Make setup case indirect to reduce memory footprint of FrameBody fomr 96 bytes to 41 bytes
Setup is only received/send once for each connection. By making it indirect, SetupBody does not need to be stored inline and because it is the biggest type by fare, it reduces the memory footprint of FrameBody significantly. This also reduces ARC traffic for a FrameBody which contains a SetupFrameBody from 5 retain/release calls to just 1.

Signed-off-by: David Nadoba <dnadoba@gmail.com>